% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_risk.R
\name{tidy.matchatr_absolute_risk}
\alias{tidy.matchatr_absolute_risk}
\title{Tidy a matchatr_absolute_risk object}
\usage{
\method{tidy}{matchatr_absolute_risk}(x, ...)
}
\arguments{
\item{x}{A \code{matchatr_absolute_risk} object.}

\item{...}{Unused.}
}
\value{
A \code{data.table}.
}
\description{
Returns the estimates table as a \code{data.table} with one row per (covariate
pattern row, evaluation time), with columns \code{row}, \code{time}, \code{estimate},
\code{ci_lower}, \code{ci_upper}.
}
\seealso{
\code{\link[=absolute_risk]{absolute_risk()}}

Other contrasts: 
\code{\link{absolute_risk}()},
\code{\link{contrast}()},
\code{\link{excess_risk}()},
\code{\link{print.matchatr_absolute_risk}()},
\code{\link{print.matchatr_excess_risk}()},
\code{\link{tidy.matchatr_excess_risk}()}
}
\concept{contrasts}
