% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_risk.R
\name{tidy.matchatr_excess_risk}
\alias{tidy.matchatr_excess_risk}
\title{Tidy a matchatr_excess_risk object}
\usage{
\method{tidy}{matchatr_excess_risk}(x, ...)
}
\arguments{
\item{x}{A \code{matchatr_excess_risk} object.}

\item{...}{Unused.}
}
\value{
A \code{data.table}.
}
\description{
Returns the estimates table as a \code{data.table} with one row per (covariate
term, evaluation time): columns \code{term}, \code{time}, \code{estimate}, \code{se}, \code{ci_lower},
\code{ci_upper}.
}
\seealso{
\code{\link[=excess_risk]{excess_risk()}}

Other contrasts: 
\code{\link{absolute_risk}()},
\code{\link{contrast}()},
\code{\link{excess_risk}()},
\code{\link{print.matchatr_absolute_risk}()},
\code{\link{print.matchatr_excess_risk}()},
\code{\link{tidy.matchatr_absolute_risk}()}
}
\concept{contrasts}
